+++ source/gwork/source/Controls/HorizontalScrollBar.cpp
@@ -97,8 +97,8 @@ void HorizontalScrollBar::OnMouseClickLeft(int x, int y, bool bDown)
 
 float HorizontalScrollBar::CalculateScrolledAmount()
 {
-    return (float)(m_bar->X()-
-                   GetButtonSize())/(float)(Width()-m_bar->Width()-(GetButtonSize()*2));
+    return static_cast<float>(m_bar->X() - GetButtonSize())
+                                / (Width() - m_bar->Width() - GetButtonSize()*2);
 }
 
 bool HorizontalScrollBar::SetScrolledAmount(float amount, bool forceUpdate)
