+++ source/gwork/source/Controls/Text.cpp
@@ -235,7 +235,7 @@ void Text::SplitWords(const Gwk::String& s, std::vector<Gwk::String>& elems)
     int w = GetParent()->Width()
                 - (GetParent()->GetPadding().left + GetParent()->GetPadding().right);
     
-    for (int i = 0; i < (int)s.length(); i++)
+    for (size_t i = 0; i < s.length(); i++)
     {
         if (s[i] == '\n')
         {
@@ -368,7 +368,7 @@ void Text::RefreshSizeWrap()
 
 unsigned int Text::NumLines()
 {
-    return (unsigned int)m_lines.size();
+    return static_cast<unsigned int>(m_lines.size());
 }
 
 Text* Text::GetLine(int i)
